+++ lib/src/widgets/formy_group_visibility.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_formy/flutter_formy.dart';
 
-class FormySubGroupVisibility extends GroupSelector<bool> {
-  FormySubGroupVisibility(
+class FormyGroupVisibility extends GroupSelector<bool> {
+  FormyGroupVisibility(
       {super.key, required super.control, required Widget child})
       : super(
           selector: (value) => value.state.isEnabled,