+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.0
+
+### Added
+- FormySubmitButton to react to GroupController's state changes
+### Fixed
+- Change FormySubGroupVisibility name to FormyGroupVisibility
+
 ## 0.0.3
 ### Fixed
 - Remove hasFocus from FieldState